Updated the create, patch, and delete review endpoints so that it automatically updates the related fields in the service and provider entity. Additionally, concurrent updates are ensured during each transaction.

+++ Server/database/migrations/sqls/20221130122423-init-tables-up.sql
@@ -127,7 +127,7 @@ CREATE TABLE IF NOT EXISTS `Review`(
     `service_id` VARCHAR(14) NOT NULL,
     `seeker_id` VARCHAR(14) NOT NULL,
     `date_timestamp` BIGINT NOT NULL,
-    `rating` FLOAT NOT NULL,
+    `rating` INT NOT NULL,
     `comment` TEXT,
     `images` TEXT,
     PRIMARY KEY (`review_id`),
